+++ src/main/java/me/zeroeightsix/kami/KamiMod.java
@@ -13,11 +13,9 @@ import me.zeroeightsix.kami.event.events.RenderEvent;
 import me.zeroeightsix.kami.event.events.RenderHudEvent;
 import me.zeroeightsix.kami.event.events.TickEvent;
 import me.zeroeightsix.kami.gui.KamiGuiScreen;
-import me.zeroeightsix.kami.gui.KamiHud;
-import me.zeroeightsix.kami.gui.windows.KamiSettings;
-import me.zeroeightsix.kami.module.Feature;
-import me.zeroeightsix.kami.module.FeatureManager;
-import me.zeroeightsix.kami.module.Listening;
+import me.zeroeightsix.kami.feature.Feature;
+import me.zeroeightsix.kami.feature.FeatureManager;
+import me.zeroeightsix.kami.feature.Listening;
 import me.zeroeightsix.kami.setting.SettingsRegister;
 import me.zeroeightsix.kami.setting.config.Configuration;
 import me.zeroeightsix.kami.util.Friends;
